Moving warehouse code into separate files
