+++ warehouse/warehouse.go
@@ -1,8 +1,6 @@
 package warehouse
 
 import (
-	"bufio"
-	"compress/gzip"
 	"database/sql"
 	"encoding/json"
 	"errors"
@@ -10,9 +8,6 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
-	"os"
-	"path/filepath"
-	"reflect"
 	"runtime"
 	"sort"
 	"strconv"
@@ -30,8 +25,6 @@ import (
 	"github.com/rudderlabs/rudder-server/rruntime"
 	"github.com/rudderlabs/rudder-server/services/db"
 	destinationConnectionTester "github.com/rudderlabs/rudder-server/services/destination-connection-tester"
-	destinationdebugger "github.com/rudderlabs/rudder-server/services/destination-debugger"
-	"github.com/rudderlabs/rudder-server/services/filemanager"
 	"github.com/rudderlabs/rudder-server/services/pgnotifier"
 	migrator "github.com/rudderlabs/rudder-server/services/sql-migrator"
 	"github.com/rudderlabs/rudder-server/services/stats"
@@ -362,26 +355,6 @@ func (wh *HandleT) backendConfigSubscriber() {
 	}
 }
 
-// syncLiveWarehouseStatus fetch last 10 records order by updated_at desc and sends uploadIds in reverse order to recordDeliveryStatus.
-// that way we can fetch last 10 records order by updated_at asc
-func (wh *HandleT) syncLiveWarehouseStatus(sourceID string, destinationID string) {
-	rows, err := wh.dbHandle.Query(fmt.Sprintf(`select id from %s where source_id='%s' and destination_id='%s' order by updated_at desc limit %d`, warehouseutils.WarehouseUploadsTable, sourceID, destinationID, warehouseSyncPreFetchCount))
-	if err != nil && err != sql.ErrNoRows {
-		panic(err)
-	}
-	defer rows.Close()
-	var uploadIDs []int64
-	for rows.Next() {
-		var uploadID int64
-		rows.Scan(&uploadID)
-		uploadIDs = append(uploadIDs, uploadID)
-	}
-
-	for index := range uploadIDs {
-		wh.recordDeliveryStatus(destinationID, uploadIDs[len(uploadIDs)-1-index])
-	}
-}
-
 // getNamespace sets namespace name in the following order
 // 	1. user set name from destinationConfig
 // 	2. from existing record in wh_schemas with same source + dest combo
@@ -464,119 +437,6 @@ func (wh *HandleT) getPendingStagingFiles(warehouse warehouseutils.WarehouseT) (
 	}
 
 	return stagingFilesList, nil
-}
-
-func (wh *HandleT) mergeSchema(currentSchema map[string]map[string]string, schemaList []map[string]map[string]string, currentMergedSchema map[string]map[string]string) map[string]map[string]string {
-	if len(currentMergedSchema) == 0 {
-		currentMergedSchema = make(map[string]map[string]string)
-	}
-	for _, schema := range schemaList {
-		for tableName, columnMap := range schema {
-			if currentMergedSchema[tableName] == nil {
-				currentMergedSchema[tableName] = make(map[string]string)
-			}
-			for columnName, columnType := range columnMap {
-				// if column already has a type in db, use that
-				if len(currentSchema) > 0 {
-					if _, ok := currentSchema[tableName]; ok {
-						if columnTypeInDB, ok := currentSchema[tableName][columnName]; ok {
-							if columnTypeInDB == "string" && columnType == "text" {
-								currentMergedSchema[tableName][columnName] = columnType
-								continue
-							}
-							currentMergedSchema[tableName][columnName] = columnTypeInDB
-							continue
-						}
-					}
-				}
-				// check if we already set the columnType in currentMergedSchema
-				if _, ok := currentMergedSchema[tableName][columnName]; !ok {
-					currentMergedSchema[tableName][columnName] = columnType
-				}
-			}
-		}
-	}
-	return currentMergedSchema
-}
-
-func (wh *HandleT) consolidateSchema(warehouse warehouseutils.WarehouseT, jsonUploadsList []*StagingFileT) map[string]map[string]string {
-	schemaInDB, err := warehouseutils.GetCurrentSchema(wh.dbHandle, warehouse)
-	if err != nil {
-		panic(err)
-	}
-
-	consolidatedSchema := make(map[string]map[string]string)
-	count := 0
-	for {
-		lastIndex := count + stagingFilesSchemaPaginationSize
-		if lastIndex >= len(jsonUploadsList) {
-			lastIndex = len(jsonUploadsList)
-		}
-
-		var ids []int64
-		for _, upload := range jsonUploadsList[count:lastIndex] {
-			ids = append(ids, upload.ID)
-		}
-
-		sqlStatement := fmt.Sprintf(`SELECT schema FROM %s WHERE id IN (%s)`, warehouseutils.WarehouseStagingFilesTable, misc.IntArrayToString(ids, ","))
-		rows, err := wh.dbHandle.Query(sqlStatement)
-		if err != nil && err != sql.ErrNoRows {
-			panic(err)
-		}
-
-		var schemas []map[string]map[string]string
-		for rows.Next() {
-			var s json.RawMessage
-			err := rows.Scan(&s)
-			if err != nil {
-				panic(err)
-			}
-			var schema map[string]map[string]string
-			err = json.Unmarshal(s, &schema)
-			if err != nil {
-				panic(err)
-			}
-
-			schemas = append(schemas, schema)
-		}
-		rows.Close()
-
-		consolidatedSchema = wh.mergeSchema(schemaInDB, schemas, consolidatedSchema)
-
-		count += stagingFilesSchemaPaginationSize
-		if count >= len(jsonUploadsList) {
-			break
-		}
-	}
-
-	// add rudder_discards table
-	destType := warehouse.Destination.DestinationDefinition.Name
-	discards := map[string]string{
-		warehouseutils.ToProviderCase(destType, "table_name"):   "string",
-		warehouseutils.ToProviderCase(destType, "row_id"):       "string",
-		warehouseutils.ToProviderCase(destType, "column_name"):  "string",
-		warehouseutils.ToProviderCase(destType, "column_value"): "string",
-		warehouseutils.ToProviderCase(destType, "received_at"):  "datetime",
-		warehouseutils.ToProviderCase(destType, "uuid_ts"):      "datetime",
-	}
-	// add loaded_at for bq to be segment compatible
-	if destType == "BQ" {
-		discards[warehouseutils.ToProviderCase(destType, "loaded_at")] = "datetime"
-	}
-	consolidatedSchema[warehouseutils.ToProviderCase(destType, warehouseutils.DiscardsTable)] = discards
-
-	// add rudder_identity_mappings table
-	if warehouseutils.IDResolutionEnabled() && misc.ContainsString(warehouseutils.IdentityEnabledWarehouses, wh.destType) {
-		identityMappings := map[string]string{
-			warehouseutils.ToProviderCase(destType, "merge_property_type"):  "string",
-			warehouseutils.ToProviderCase(destType, "merge_property_value"): "string",
-			warehouseutils.ToProviderCase(destType, "rudder_id"):            "string",
-			warehouseutils.ToProviderCase(destType, "updated_at"):           "datetime",
-		}
-		consolidatedSchema[warehouseutils.ToProviderCase(destType, warehouseutils.IdentityMappingsTable)] = identityMappings
-	}
-
-	return consolidatedSchema
 }
 
 func (wh *HandleT) areTableUploadsCreated(upload warehouseutils.UploadT) bool {
@@ -767,22 +627,6 @@ func isDestInProgress(warehouse warehouseutils.WarehouseT) bool {
 	return false
 }
 
-func isDestPreLoaded(warehouse warehouseutils.WarehouseT) bool {
-	preLoadedIdentitiesMapLock.RLock()
-	if preLoadedIdentitiesMap[connectionString(warehouse)] {
-		preLoadedIdentitiesMapLock.RUnlock()
-		return true
-	}
-	preLoadedIdentitiesMapLock.RUnlock()
-	return false
-}
-
-func setDestPreLoaded(warehouse warehouseutils.WarehouseT) {
-	preLoadedIdentitiesMapLock.Lock()
-	preLoadedIdentitiesMap[connectionString(warehouse)] = true
-	preLoadedIdentitiesMapLock.Unlock()
-}
-
 func uploadFrequencyExceeded(warehouse warehouseutils.WarehouseT, syncFrequency string) bool {
 	freqInS := uploadFreqInS
 	if syncFrequency != "" {
@@ -801,177 +645,6 @@ func setLastExec(warehouse warehouseutils.WarehouseT) {
 	lastExecMapLock.Lock()
 	defer lastExecMapLock.Unlock()
 	lastExecMap[connectionString(warehouse)] = timeutil.Now().Unix()
-}
-
-func (wh *HandleT) hasLocalIdentityData(warehouse warehouseutils.WarehouseT) bool {
-	sqlStatement := fmt.Sprintf(`SELECT count(*) FROM %s`, warehouseutils.IdentityMergeRulesTableName(warehouse))
-	var count int
-	err := wh.dbHandle.QueryRow(sqlStatement).Scan(&count)
-	if err != nil {
-		// TOOD: Handle this
-		panic(err)
-	}
-	return count > 0
-}
-
-func (wh *HandleT) hasWarehouseData(warehouse warehouseutils.WarehouseT) (bool, error) {
-	whManager, err := warehousemanager.NewWhManager(wh.destType)
-	if err != nil {
-		panic(err)
-	}
-
-	empty, err := whManager.IsEmpty(warehouse)
-	if err != nil {
-		return false, err
-	}
-	return !empty, nil
-
-	// TODO: Change logic to check if warehouse has data in tables
-	// sqlStatement := fmt.Sprintf(`SELECT count(*) FROM %s WHERE destination_id='%s' AND status='%s'`, warehouseutils.WarehouseUploadsTable, warehouse.Destination.ID, warehouseutils.ExportedDataState)
-	// var count int
-	// err := wh.dbHandle.QueryRow(sqlStatement).Scan(&count)
-	// if err != nil {
-	// 	// TOOD: Handle this
-	// 	panic(err)
-	// }
-	// return count > 0
-}
-
-func (wh *HandleT) setupIdentityTables(warehouse warehouseutils.WarehouseT) {
-	var name sql.NullString
-	sqlStatement := fmt.Sprintf(`SELECT to_regclass('%s')`, warehouseutils.IdentityMappingsTableName(warehouse))
-	err := wh.dbHandle.QueryRow(sqlStatement).Scan(&name)
-	if err != nil {
-		panic(err)
-	}
-	if len(name.String) > 0 {
-		return
-	}
-	// create tables
-
-	sqlStatement = fmt.Sprintf(`
-		CREATE TABLE IF NOT EXISTS %s (
-			id BIGSERIAL PRIMARY KEY,
-			merge_property_1_type VARCHAR(64) NOT NULL,
-			merge_property_1_value TEXT NOT NULL,
-			merge_property_2_type VARCHAR(64),
-			merge_property_2_value TEXT,
-			created_at TIMESTAMP NOT NULL DEFAULT NOW());
-		`, warehouseutils.IdentityMergeRulesTableName(warehouse),
-	)
-
-	_, err = wh.dbHandle.Exec(sqlStatement)
-	if err != nil {
-		panic(err)
-	}
-
-	sqlStatement = fmt.Sprintf(`
-		CREATE TABLE IF NOT EXISTS %s (
-			id BIGSERIAL PRIMARY KEY,
-			merge_property_type VARCHAR(64) NOT NULL,
-			merge_property_value TEXT NOT NULL,
-			rudder_id VARCHAR(64) NOT NULL,
-			updated_at TIMESTAMP NOT NULL DEFAULT NOW());
-		`, warehouseutils.IdentityMappingsTableName(warehouse),
-	)
-
-	_, err = wh.dbHandle.Exec(sqlStatement)
-	if err != nil {
-		panic(err)
-	}
-
-	sqlStatement = fmt.Sprintf(`
-		ALTER TABLE %s
-			ADD CONSTRAINT unique_merge_property_%s UNIQUE (merge_property_type, merge_property_value);
-		`, warehouseutils.IdentityMappingsTableName(warehouse), warehouse.Destination.ID,
-	)
-
-	_, err = wh.dbHandle.Exec(sqlStatement)
-	if err != nil {
-		panic(err)
-	}
-}
-
-func (wh *HandleT) initPreLoadUpload(warehouse warehouseutils.WarehouseT) warehouseutils.UploadT {
-	schema := make(map[string]map[string]string)
-	// TODO: DRY this code
-	identityRules := map[string]string{
-		warehouseutils.ToProviderCase(wh.destType, "merge_property_1_type"):  "string",
-		warehouseutils.ToProviderCase(wh.destType, "merge_property_1_value"): "string",
-		warehouseutils.ToProviderCase(wh.destType, "merge_property_2_type"):  "string",
-		warehouseutils.ToProviderCase(wh.destType, "merge_property_2_value"): "string",
-	}
-	schema[warehouseutils.ToProviderCase(wh.destType, warehouseutils.IdentityMergeRulesTable)] = identityRules
-
-	// add rudder_identity_mappings table
-	identityMappings := map[string]string{
-		warehouseutils.ToProviderCase(wh.destType, "merge_property_type"):  "string",
-		warehouseutils.ToProviderCase(wh.destType, "merge_property_value"): "string",
-		warehouseutils.ToProviderCase(wh.destType, "rudder_id"):            "string",
-		warehouseutils.ToProviderCase(wh.destType, "updated_at"):           "datetime",
-	}
-	schema[warehouseutils.ToProviderCase(wh.destType, warehouseutils.IdentityMappingsTable)] = identityMappings
-
-	marshalledSchema, err := json.Marshal(schema)
-	if err != nil {
-		panic(err)
-	}
-
-	sqlStatement := fmt.Sprintf(`INSERT INTO %s (source_id, namespace, destination_id, destination_type, status, schema, error, created_at, updated_at, start_staging_file_id, end_staging_file_id, start_load_file_id, end_load_file_id)	VALUES ($1, $2, $3, $4, $5, $6 ,$7, $8, $9, $10, $11, $12, $13) RETURNING id`, warehouseutils.WarehouseUploadsTable)
-	stmt, err := wh.dbHandle.Prepare(sqlStatement)
-	if err != nil {
-		panic(err)
-	}
-	defer stmt.Close()
-
-	now := timeutil.Now()
-	row := stmt.QueryRow(warehouse.Source.ID, warehouse.Namespace, warehouse.Destination.ID, wh.preLoadDestType(), warehouseutils.WaitingState, marshalledSchema, "{}", now, now, 0, 0, 0, 0)
-
-	var uploadID int64
-	err = row.Scan(&uploadID)
-	if err != nil {
-		panic(err)
-	}
-
-	upload := warehouseutils.UploadT{
-		ID:              uploadID,
-		Namespace:       warehouse.Namespace,
-		SourceID:        warehouse.Source.ID,
-		DestinationID:   warehouse.Destination.ID,
-		DestinationType: wh.preLoadDestType(),
-		Status:          warehouseutils.WaitingState,
-		Schema:          schema,
-	}
-
-	err = wh.initTableUploads(upload, upload.Schema)
-	if err != nil {
-		panic(err)
-	}
-
-	return upload
-}
-
-func (wh *HandleT) preLoadIdentityTables(warehouse warehouseutils.WarehouseT) (upload warehouseutils.UploadT, err error) {
-
-	// check for pending preLoads
-	var found bool
-	if upload, found = wh.getPendingPreLoad(warehouse); !found {
-		upload = wh.initPreLoadUpload(warehouse)
-	}
-
-	whManager, err := warehousemanager.NewWhManager(wh.destType)
-	if err != nil {
-		panic(err)
-	}
-
-	err = whManager.Process(warehouseutils.ConfigT{
-		DbHandle:  wh.dbHandle,
-		Upload:    upload,
-		Warehouse: warehouse,
-		Stage:     warehouseutils.PreLoadingIdentities,
-	})
-
-	return
 }
 
 func (wh *HandleT) mainLoop() {
@@ -1145,16 +818,6 @@ type PayloadT struct {
 	LoadFileIDs         []int64
 }
 
-// Temporary store for processing staging file to load file
-type JobRunT struct {
-	job                  PayloadT
-	stagingFilePath      string
-	uuidTS               time.Time
-	outputFileWritersMap map[string]misc.GZipWriter
-	tableEventCountMap   map[string]int
-	stagingFileReader    *gzip.Reader
-}
-
 func (wh *HandleT) createLoadFiles(job *ProcessStagingFilesJobT) (err error) {
 	// stat for time taken to process staging files in a single job
 	timer := warehouseutils.DestStat(stats.TimerType, "process_staging_files_batch_time", job.Warehouse.Destination.ID)
@@ -1300,87 +963,6 @@ func (wh *HandleT) SyncLoadFilesToWarehouse(job *ProcessStagingFilesJobT) (err e
 	return
 }
 
-func (wh *HandleT) recordDeliveryStatus(destID string, uploadID int64) {
-	if !destinationdebugger.HasUploadEnabled(destID) {
-		return
-	}
-	var (
-		sourceID      string
-		destinationID string
-		status        string
-		errorCode     string
-		errorResp     string
-		updatedAt     time.Time
-		tableName     string
-		tableStatus   string
-		attemptNum    int
-	)
-	successfulTableUploads := make([]string, 0)
-	failedTableUploads := make([]string, 0)
-
-	row := wh.dbHandle.QueryRow(fmt.Sprintf(`select source_id, destination_id, status, error, updated_at from %s where id=%d`, warehouseutils.WarehouseUploadsTable, uploadID))
-	err := row.Scan(&sourceID, &destinationID, &status, &errorResp, &updatedAt)
-	if err != nil && err != sql.ErrNoRows {
-		panic(err)
-	}
-	rows, err := wh.dbHandle.Query(fmt.Sprintf(`select table_name, status from %s where wh_upload_id=%d`, warehouseutils.WarehouseTableUploadsTable, uploadID))
-	if err != nil && err != sql.ErrNoRows {
-		panic(err)
-	}
-	defer rows.Close()
-	for rows.Next() {
-		rows.Scan(&tableName, &tableStatus)
-		if tableStatus == warehouseutils.ExportedDataState {
-			successfulTableUploads = append(successfulTableUploads, tableName)
-		} else {
-			failedTableUploads = append(failedTableUploads, tableName)
-		}
-	}
-
-	var errJSON map[string]map[string]interface{}
-	err = json.Unmarshal([]byte(errorResp), &errJSON)
-	if err != nil {
-		panic(err)
-	}
-	if stateErr, ok := errJSON[status]; ok {
-		if attempt, ok := stateErr["attempt"]; ok {
-			if floatAttempt, ok := attempt.(float64); ok {
-				attemptNum = attemptNum + int(floatAttempt)
-			}
-		}
-	}
-	if attemptNum == 0 {
-		attemptNum = 1
-	}
-	var errorRespB []byte
-	if errorResp == "{}" {
-		errorCode = "200"
-	} else {
-		errorCode = "400"
-
-	}
-	errorRespB, _ = json.Marshal(ErrorResponseT{Error: errorResp})
-
-	payloadMap := map[string]interface{}{
-		"lastSyncedAt":             updatedAt,
-		"successful_table_uploads": successfulTableUploads,
-		"failed_table_uploads":     failedTableUploads,
-		"uploadID":                 uploadID,
-		"error":                    errorResp,
-	}
-	payload, _ := json.Marshal(payloadMap)
-	deliveryStatus := destinationdebugger.DeliveryStatusT{
-		DestinationID: destinationID,
-		SourceID:      sourceID,
-		Payload:       payload,
-		AttemptNum:    attemptNum,
-		JobState:      status,
-		ErrorCode:     errorCode,
-		ErrorResponse: errorRespB,
-	}
-	destinationdebugger.RecordEventDeliveryStatus(destinationID, &deliveryStatus)
-}
-
 func getBucketFolder(batchID string, tableName string) string {
 	return fmt.Sprintf(`%v-%v`, batchID, tableName)
 }
@@ -1441,374 +1023,6 @@ var loadFileFormatMap = map[string]string{
 	"SNOWFLAKE":  "csv",
 	"POSTGRES":   "csv",
 	"CLICKHOUSE": "csv",
-}
-
-func (jobRun *JobRunT) setStagingFileReader() (reader *gzip.Reader, endOfFile bool) {
-
-	job := jobRun.job
-	logger.Debugf("Starting read from downloaded staging file: %s", job.StagingFileLocation)
-	rawf, err := os.Open(jobRun.stagingFilePath)
-	if err != nil {
-		logger.Errorf("[WH]: Error opening file using os.Open at path:%s downloaded from %s", jobRun.stagingFilePath, job.StagingFileLocation)
-		panic(err)
-	}
-	reader, err = gzip.NewReader(rawf)
-	if err != nil {
-		if err.Error() == "EOF" {
-			return nil, true
-		}
-		logger.Errorf("[WH]: Error reading file using gzip.NewReader at path:%s downloaded from %s", jobRun.stagingFilePath, job.StagingFileLocation)
-		panic(err)
-	}
-
-	jobRun.stagingFileReader = reader
-	return reader, false
-}
-
-/*
- * Get download path for the job. Also creates missing directories for this path
- */
-func (jobRun *JobRunT) setStagingFileDownloadPath() (filePath string) {
-	job := jobRun.job
-	dirName := "/rudder-warehouse-json-uploads-tmp/"
-	tmpDirPath, err := misc.CreateTMPDIR()
-	if err != nil {
-		logger.Errorf("[WH]: Failed to create tmp DIR")
-		panic(err)
-	}
-	filePath = tmpDirPath + dirName + fmt.Sprintf(`%s_%s/`, job.DestinationType, job.DestinationID) + job.StagingFileLocation
-	err = os.MkdirAll(filepath.Dir(filePath), os.ModePerm)
-	if err != nil {
-		panic(err)
-	}
-	jobRun.stagingFilePath = filePath
-	return filePath
-}
-
-// Get fileManager
-func (job *PayloadT) getFileManager() (filemanager.FileManager, error) {
-	fileManager, err := filemanager.New(&filemanager.SettingsT{
-		Provider: warehouseutils.ObjectStorageType(job.DestinationType, job.DestinationConfig),
-		Config:   misc.GetObjectStorageConfig(warehouseutils.ObjectStorageType(job.DestinationType, job.DestinationConfig), job.DestinationConfig),
-	})
-	return fileManager, err
-}
-
-/*
- * Download Staging file for the job
- */
-func (jobRun *JobRunT) downloadStagingFile() error {
-	filePath := jobRun.stagingFilePath
-	file, err := os.Create(filePath)
-	if err != nil {
-		panic(err)
-	}
-
-	job := jobRun.job
-	downloader, err := jobRun.job.getFileManager()
-	if err != nil {
-		logger.Errorf("[WH]: Failed to initialize downloader")
-		return err
-	}
-
-	timer := warehouseutils.DestStat(stats.TimerType, "download_staging_file_time", job.DestinationID)
-	timer.Start()
-
-	err = downloader.Download(file, job.StagingFileLocation)
-	if err != nil {
-		logger.Errorf("[WH]: Failed to download file")
-		return err
-	}
-	file.Close()
-	timer.End()
-
-	fi, err := os.Stat(filePath)
-	if err != nil {
-		logger.Errorf("[WH]: Error getting file size of downloaded staging file: ", err)
-	}
-	fileSize := fi.Size()
-	logger.Debugf("[WH]: Downloaded staging file %s size:%v", job.StagingFileLocation, fileSize)
-	warehouseutils.DestStat(stats.CountType, "downloaded_staging_file_size", job.DestinationID).Count(int(fileSize))
-
-	return nil
-}
-
-func (job *PayloadT) getDiscardsTable() string {
-	return warehouseutils.ToProviderCase(job.DestinationType, warehouseutils.DiscardsTable)
-}
-
-func (jobRun *JobRunT) getLoadFilePath(tableName string) string {
-	job := jobRun.job
-	return strings.TrimSuffix(jobRun.stagingFilePath, "json.gz") + tableName + fmt.Sprintf(`.%s`, loadFileFormatMap[job.DestinationType]) + ".gz"
-}
-
-func (job *PayloadT) getColumnName(columnName string) string {
-	return warehouseutils.ToProviderCase(job.DestinationType, columnName)
-}
-
-func (jobRun *JobRunT) uploadLoadFileToObjectStorage(uploader filemanager.FileManager, uploadFile *misc.GZipWriter, tableName string) (filemanager.UploadOutput, error) {
-	job := jobRun.job
-	uploadFile.CloseGZ()
-	file, err := os.Open(uploadFile.File.Name())
-	if err != nil {
-		logger.Errorf("[WH]: Failed to Open File: %s", uploadFile.File.Name())
-		return filemanager.UploadOutput{}, err
-	}
-
-	defer os.Remove(uploadFile.File.Name())
-	logger.Debugf("[WH]: %s: Uploading load_file to %s for table: %s in staging_file id: %v", job.DestinationType, warehouseutils.ObjectStorageType(job.DestinationType, job.DestinationConfig), tableName, job.StagingFileID)
-	uploadLocation, err := uploader.Upload(file, config.GetEnv("WAREHOUSE_BUCKET_LOAD_OBJECTS_FOLDER_NAME", "rudder-warehouse-load-objects"), tableName, job.SourceID, getBucketFolder(job.BatchID, tableName))
-	return uploadLocation, err
-}
-
-func (job *PayloadT) markLoadFileUploadSuccess(tableName string, uploadLocation string, numEvents int) int64 {
-	sqlStatement := fmt.Sprintf(`INSERT INTO %s (staging_file_id, location, source_id, destination_id, destination_type, table_name, total_events, created_at)
-	VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING id`, warehouseutils.WarehouseLoadFilesTable)
-	stmt, err := dbHandle.Prepare(sqlStatement)
-	if err != nil {
-		panic(err)
-	}
-	var fileID int64
-	err = stmt.QueryRow(job.StagingFileID, uploadLocation, job.SourceID, job.DestinationID, job.DestinationType, tableName, numEvents, timeutil.Now()).Scan(&fileID)
-	if err != nil {
-		panic(err)
-	}
-	stmt.Close()
-	return fileID
-}
-
-// Sort columns per table to maintain same order in load file (needed in case of csv load file)
-func (job *PayloadT) getSortedColumnMapForAllTables() map[string][]string {
-	sortedTableColumnMap := make(map[string][]string)
-	for tableName, columnMap := range job.Schema {
-		sortedTableColumnMap[tableName] = []string{}
-		for k := range columnMap {
-			sortedTableColumnMap[tableName] = append(sortedTableColumnMap[tableName], k)
-		}
-		sort.Strings(sortedTableColumnMap[tableName])
-	}
-	return sortedTableColumnMap
-}
-
-func (jobRun *JobRunT) getWriter(tableName string) (misc.GZipWriter, error) {
-	writer, ok := jobRun.outputFileWritersMap[tableName]
-	if !ok {
-		outputFilePath := jobRun.getLoadFilePath(tableName)
-		var err error
-		writer, err = misc.CreateGZ(outputFilePath)
-		if err != nil {
-			return misc.GZipWriter{}, err
-		}
-		jobRun.outputFileWritersMap[tableName] = writer
-		jobRun.tableEventCountMap[tableName] = 0
-	}
-	return writer, nil
-}
-
-func (jobRun *JobRunT) cleanup() {
-	if jobRun.stagingFileReader != nil {
-		err := jobRun.stagingFileReader.Close()
-		logger.Errorf("[WH]: Failed to close staging file: %w", err)
-	}
-
-	if jobRun.stagingFilePath != "" {
-		err := os.Remove(jobRun.stagingFilePath)
-		if err != nil {
-			logger.Errorf("[WH]: Failed to remove staging file: %w", err)
-		}
-	}
-
-	if jobRun.outputFileWritersMap != nil {
-		for _, writer := range jobRun.outputFileWritersMap {
-			err := writer.CloseGZ()
-			if err != nil {
-				logger.Errorf("[WH]: Failed to closed output load file: %w", err)
-			}
-		}
-	}
-}
-
-func (event *BatchRouterEventT) getColumnInfo(columnName string) (columnInfo ColumnInfoT, ok bool) {
-	columnVal, ok := event.Data[columnName]
-	if !ok {
-		return ColumnInfoT{}, false
-	}
-	columnType, ok := event.Metadata.Columns[columnName]
-	if !ok {
-		return ColumnInfoT{}, false
-	}
-	return ColumnInfoT{ColumnVal: columnVal, ColumnType: columnType}, true
-}
-
-//
-// This function is triggered when warehouse-master creates a new entry in wh_uploads table
-// This is executed in the context of the warehouse-slave/worker and does the following:
-//
-// 1. Download the Staging file into a tmp directory
-// 2. Transform the staging file into multiple load files (One file per output table)
-// 3. Uploads these load files to Object storage
-// 4. Save entries for the generated load files in wh_load_files table
-// 5. Delete the staging and load files from tmp directory
-//
-
-func processStagingFile(job PayloadT) (loadFileIDs []int64, err error) {
-	logger.Debugf("[WH]: Starting processing staging file: %v at %v for %s:%s", job.StagingFileID, job.StagingFileLocation, job.DestinationType, job.DestinationID)
-
-	jobRun := JobRunT{job: job}
-	defer jobRun.cleanup()
-
-	jobRun.setStagingFileDownloadPath()
-
-	// This creates the file, so on successful creation remove it
-	err = jobRun.downloadStagingFile()
-	if err != nil {
-		return loadFileIDs, err
-	}
-
-	sortedTableColumnMap := job.getSortedColumnMapForAllTables()
-
-	reader, endOfFile := jobRun.setStagingFileReader()
-	if endOfFile {
-		// If empty file, return nothing
-		return loadFileIDs, nil
-	}
-	scanner := bufio.NewScanner(reader)
-
-	// read from staging file and write a separate load file for each table in warehouse
-	jobRun.outputFileWritersMap = make(map[string]misc.GZipWriter)
-	jobRun.tableEventCountMap = make(map[string]int)
-	jobRun.uuidTS = timeutil.Now()
-	misc.PrintMemUsage()
-
-	// Initilize Discards Table
-	discardsTable := job.getDiscardsTable()
-	jobRun.tableEventCountMap[discardsTable] = 0
-
-	timer := warehouseutils.DestStat(stats.TimerType, "process_staging_file_to_csv_time", job.DestinationID)
-	timer.Start()
-
-	lineBytesCounter := 0
-	var interfaceSliceSample []interface{}
-	for scanner.Scan() {
-		lineBytes := scanner.Bytes()
-		lineBytesCounter += len(lineBytes)
-		var batchRouterEvent BatchRouterEventT
-		err := json.Unmarshal(lineBytes, &batchRouterEvent)
-		if err != nil {
-			logger.Errorf("[WH]: Failed to unmarshal JSON line to batchrouter event: %+v", batchRouterEvent)
-			continue
-		}
-
-		tableName := batchRouterEvent.Metadata.Table
-		columnData := batchRouterEvent.Data
-
-		// Create separate load file for each table
-		gzWriter, err := jobRun.getWriter(tableName)
-		if err != nil {
-			return nil, err
-		}
-
-		eventLoader := warehouseutils.GetNewEventLoader(job.DestinationType)
-		for _, columnName := range sortedTableColumnMap[tableName] {
-			if eventLoader.IsLoadTimeColumn(columnName) {
-				timestampFormat := eventLoader.GetLoadTimeFomat(columnName)
-				eventLoader.AddColumn(job.getColumnName(columnName), jobRun.uuidTS.Format(timestampFormat))
-				continue
-			}
-			columnInfo, ok := batchRouterEvent.getColumnInfo(columnName)
-			if !ok {
-				eventLoader.AddEmptyColumn(columnName)
-				continue
-			}
-			columnType := columnInfo.ColumnType
-			columnVal := columnInfo.ColumnVal
-
-			// TODO: Move this into function that handles int type
-			if columnType == "int" || columnType == "bigint" {
-				floatVal, ok := columnVal.(float64)
-				if !ok {
-					eventLoader.AddEmptyColumn(columnName)
-					continue
-				}
-				columnVal = int(floatVal)
-			}
-
-			dataTypeInSchema, ok := job.Schema[tableName][columnName]
-			if ok && columnType != dataTypeInSchema {
-				newColumnVal, ok := handleSchemaChange(dataTypeInSchema, columnType, columnVal)
-				if !ok {
-					eventLoader.AddEmptyColumn(columnName)
-
-					discardWriter, err := jobRun.getWriter(discardsTable)
-					if err != nil {
-						return nil, err
-					}
-					err = jobRun.handleDiscardTypes(tableName, columnName, columnVal, columnData, discardWriter)
-
-					if err != nil {
-						logger.Error("[WH]: Failed to write to discards: %w", err)
-					}
-					jobRun.tableEventCountMap[discardsTable]++
-					continue
-				}
-				if newColumnVal == nil {
-					eventLoader.AddEmptyColumn(columnName)
-					continue
-				}
-				columnVal = newColumnVal
-			}
-
-			// Special handling for JSON arrays
-			// TODO: Will this work for both BQ and RS?
-			if reflect.TypeOf(columnVal) == reflect.TypeOf(interfaceSliceSample) {
-				marshalledVal, err := json.Marshal(columnVal)
-				if err != nil {
-					logger.Errorf("[WH]: Error in marshalling []interface{} columnVal: %w", err)
-					eventLoader.AddEmptyColumn(columnName)
-					continue
-				}
-				columnVal = string(marshalledVal)
-			}
-
-			eventLoader.AddColumn(columnName, columnVal)
-		}
-
-		// Completed parsing all columns, write single event to the file
-		eventData, err := eventLoader.WriteToString()
-		if err != nil {
-			logger.Errorf("[WH]: Failed to write event to string: %w", err)
-			return loadFileIDs, err
-		}
-		gzWriter.WriteGZ(eventData)
-		jobRun.tableEventCountMap[tableName]++
-	}
-
-	timer.End()
-	misc.PrintMemUsage()
-
-	logger.Debugf("[WH]: Process %v bytes from downloaded staging file: %s", lineBytesCounter, job.StagingFileLocation)
-	warehouseutils.DestStat(stats.CountType, "bytes_processed_in_staging_file", job.DestinationID).Count(lineBytesCounter)
-
-	// Upload each generated load file to ObjectStorage
-	// On successful upload, store the saved fileID in wh_load_files table
-	uploader, err := job.getFileManager()
-	if err != nil {
-		panic(err)
-	}
-
-	timer = warehouseutils.DestStat(stats.TimerType, "upload_load_files_per_staging_file_time", job.DestinationID)
-	timer.Start()
-	for tableName, outputFile := range jobRun.outputFileWritersMap {
-		uploadOutput, err := jobRun.uploadLoadFileToObjectStorage(uploader, &outputFile, tableName)
-		if err != nil {
-			// TODO: If we break in between, we have only few load files. How do we handle this?
-			return loadFileIDs, err
-		}
-		fileID := job.markLoadFileUploadSuccess(tableName, uploadOutput.Location, jobRun.tableEventCountMap[tableName])
-		loadFileIDs = append(loadFileIDs, fileID)
-	}
-	timer.End()
-	return loadFileIDs, nil
 }
 
 // Gets the config from config backend and extracts enabled writekeys
